Cover legacy build graph file read effects
