+++ tests/integration/cli_build.rs
@@ -30,3 +30,5 @@ mod graph_validation_test_command_host_effects;
 mod graph_validation_test_command_validation;
 #[path = "cli_build/legacy_graph_command.rs"]
 mod legacy_graph_command;
+#[path = "cli_build/legacy_graph_command_host_effects.rs"]
+mod legacy_graph_command_host_effects;
